Add files via upload
Anagram Solver
AnagramSolver is a Java-based program that solves anagram problems. Given a set of words, the program checks if any two words are anagrams (i.e., they contain the same letters in a different order) and groups all anagrams together.


Description
This project is designed to solve the problem of finding anagrams in a list of words. It efficiently groups words that are anagrams of each other and outputs them. The program uses Java Collections, Maps, and Sorting to achieve the desired results. It also demonstrates various Java techniques for string manipulation and algorithms.

Example:
Input: ["listen", "silent", "enlist", "hello", "world"]
Output: [["listen", "silent", "enlist"], ["hello"], ["world"]]
Technologies Used
Java 1.7/1.8: The programming language used for the application.
Java Collections Framework: Used for handling the list of words and grouping the anagrams.
HashMap: Used to group words that are anagrams by sorting the characters in the word.